+++ src/libmbim-glib/mbim-message.c
@@ -235,7 +235,7 @@ _mbim_message_read_guint32_array (const MbimMessage  *self,
 
 gboolean
 _mbim_message_read_guint64 (const MbimMessage  *self,
-                            guint64             relative_offset,
+                            guint32             relative_offset,
                             guint64            *value,
                             GError            **error)
 {
